WebAdmin/requirements.txt
altgraph==0.17.4
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-htmx==1.18.0
django-steamauth==1.1.2.3
idna==3.7
mod-wsgi==5.0.0
mysqlclient==2.2.4
mysql-connector-python==9.0.0
packaging==24.1
pefile==2023.2.7
pip==22.3
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pywin32-ctypes==0.2.2
requests==2.32.3
setuptools==65.5.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
